+++ test/com/walmartlabs/lacinia/validator_test.clj
@@ -271,13 +271,14 @@
 
 (deftest query-argument-validations
     (let [q "{ echoArgs(integer: \"hello world\") { integer } }"]
-      (is (= {:errors [{:extensions {:argument :integer
-                                     :field :echoArgs
-                                     :type-name :Int
-                                     :value "hello world"}
-                        :locations [{:column 3
-                                     :line 1}]
-                        :message "Exception applying arguments to field `echoArgs': For argument `integer', scalar value is not parsable as type `Int'."}]}
+      (is (= {:errors
+              [{:extensions {:argument :integer
+                             :field :echoArgs
+                             :type-name :Int
+                             :value "hello world"}
+                :locations [{:column 3
+                             :line 1}]
+                :message "Exception applying arguments to field `echoArgs': For argument `integer', scalar value is not parsable as type `Int': Unable to to perform cooercion: For input string: \"hello world\""}]}
              (execute compiled-schema q {} nil))))
 
   (testing "undefined argument"
@@ -300,14 +301,13 @@
                                       nestedInputObject: {integerArray: \"hello world\",
                                                           date: \"1983-08-13\"}}) { integer }}"]
       (is (= {:errors
-              [{:message
-                "Exception applying arguments to field `echoArgs': For argument `inputObject', scalar value is not parsable as type `Int'.",
-                :locations [{:line 1
-                             :column 3}]
-                :extensions {:field :echoArgs
-                             :argument :inputObject
-                             :value "hello world"
-                             :type-name :Int}}]}
+              [{:extensions {:argument :inputObject
+                             :field :echoArgs
+                             :type-name :Int
+                             :value "hello world"}
+                :locations [{:column 3
+                             :line 1}]
+                :message "Exception applying arguments to field `echoArgs': For argument `inputObject', scalar value is not parsable as type `Int': Unable to to perform cooercion: For input string: \"hello world\""}]}
 
              (execute compiled-schema q {} nil)))))
 
@@ -334,13 +334,15 @@
 
   (testing "invalid array element"
     (let [q "{echoArgs(integer: 3, integerArray: [1, 2, \"foo\"]) { integer }}"]
-      (is (= {:errors [{:extensions {:argument :integerArray
-                                     :field :echoArgs
-                                     :type-name :Int
-                                     :value "foo"}
-                        :locations [{:column 2
-                                     :line 1}]
-                        :message "Exception applying arguments to field `echoArgs': For argument `integerArray', scalar value is not parsable as type `Int'."}]}
+      (is (= {:errors
+              [{:extensions {:argument :integerArray
+                             :field :echoArgs
+                             :type-name :Int
+                             :value "foo"}
+                :locations [{:column 2
+                             :line 1}]
+                :message "Exception applying arguments to field `echoArgs': For argument `integerArray', scalar value is not parsable as type `Int': Unable to to perform cooercion: For input string: \"foo\""}]}
+
              (execute compiled-schema q {} nil)))))
 
   (testing "valid arguments"